ai_agent/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
gunicorn==20.1.0
python-dotenv==0.21.0
google-generativeai==0.4.0
langchain==0.1.5
